requirements.txt
customtkinter
threading
websockets
pycryptodome